hardcoding tracer error so I can move forward

+++ python/jumanji_env/environments/complex_orchard/generator.py
@@ -78,11 +78,7 @@ class ComplexOrchardGenerator:
 
     The final return is a flat array of the diameter of each tree.
     """
-    
-    # Ensure concrete inputs for tree distances
-    tree_row_distance = tree_row_distance.astype(float)
-    tree_col_distance = tree_col_distance.astype(float)
-    
+  
     tree_x = jnp.arange(tree_row_distance, self.width - tree_row_distance, tree_row_distance / 2)
     tree_y = jnp.arange(tree_col_distance * 2, self.height - tree_col_distance / 2, tree_col_distance / 2)
     position_offset = self.random_normal(key, (len(tree_x), len(tree_y), 2), TREE_VARIATION)
